requirements.txt
anyio==4.5.0
async-property==0.2.2
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
deprecation==2.1.0
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
jwt==1.3.1
logging==0.4.9.6
MarkupSafe==2.1.5
packaging==24.1
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python-keycloak==4.4.0
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
